% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{backup_to_cloud}
\alias{backup_to_cloud}
\title{Backup to cloud}
\usage{
backup_to_cloud(path = "commuteR")
}
\arguments{
\item{path}{a folder in Dropbox where the backup files are to be saved (typically, the same \code{path} given to \code{restore_from_cloud()}}
}
\description{
Backup to cloud
}
