% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicago_to_harvard_monday_through_friday.R
\docType{data}
\name{chicago_to_harvard_monday_through_friday}
\alias{chicago_to_harvard_monday_through_friday}
\title{Select train times from Ogilvie to Arlington Heights.}
\format{A data frame with 7 rows and 3 fields:
\describe{
  \item{train_number}{train number}
  \item{Ogilvie}{time train departs}
  \item{Arlington Heights}{time train arrives}
}}
\source{
\url{https://metrarail.com/}
}
\usage{
chicago_to_harvard_monday_through_friday
}
\description{
Select train times from Ogilvie to Arlington Heights.
}
\keyword{datasets}
