% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{filter_installed_packages}
\alias{filter_installed_packages}
\title{Filter package names of packages that already installed}
\usage{
filter_installed_packages(packages)
}
\arguments{
\item{packages}{A string vector of names of packages.}
}
\value{
\code{packages}, without any packages that already installed (any version).
}
\description{
Filter package names of packages that already installed
}
