% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvard_to_chicago_monday_through_friday.R
\docType{data}
\name{harvard_to_chicago_monday_through_friday}
\alias{harvard_to_chicago_monday_through_friday}
\title{Select train times from Arlington Heights to Ogilvie.}
\format{A data frame with 7 rows and 3 fields:
\describe{
  \item{train_number}{train number}
  \item{Arlington Heights}{time train departs}
  \item{Ogilvie}{time train arrives}
}}
\source{
\url{https://metrarail.com/}
}
\usage{
harvard_to_chicago_monday_through_friday
}
\description{
Select train times from Arlington Heights to Ogilvie.
}
\keyword{datasets}
