% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{install_github_packages}
\alias{install_github_packages}
\title{Install packages fom GitHub}
\usage{
install_github_packages(packages, dependencies = NA, quiet = T,
  force = F, lib = .libPaths()[1])
}
\arguments{
\item{packages}{A string vector of names of packages (e.g. "dplyr").}

\item{dependencies}{logical indicating whether to also install uninstalled packages which these packages depend on/link to/import/suggest (and so on recursively). Not used if repos = NULL. Can also be a character vector, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances").
Only supported if lib is of length one (or missing), so it is unambiguous where to install the dependent packages. If this is not the case it is ignored, with a warning.
The default, NA, means c("Depends", "Imports", "LinkingTo").
TRUE means to use c("Depends", "Imports", "LinkingTo", "Suggests") for pkgs and c("Depends", "Imports", "LinkingTo") for added dependencies: this installs all the packages needed to run pkgs, their examples, tests and vignettes (if the package author specified them correctly).
In all of these, "LinkingTo" is omitted for binary packages.}

\item{quiet}{logical: if true, reduce the amount of output.}

\item{force}{logical. Should package be installed if it already exists on local system?}

\item{lib}{character vector giving the library directories where to install the packages. Recycled as needed. If missing, defaults to the first element of .libPaths().}
}
\description{
Install packages fom GitHub
}
