% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{install_packages_from_backup}
\alias{install_packages_from_backup}
\title{Install packages, from both CRAN and GitHub}
\usage{
install_packages_from_backup(cran_packages, gh_packages)
}
\arguments{
\item{cran_packages}{A string vector with names of packages to be installed from CRAN}

\item{gh_packages}{A string vector with names of repo\\packages to be installed from GitHub}
}
\description{
Installs packages, from both CRAN and GitHub, and skips any packages that are already installed
}
