% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{list_backup}
\alias{list_backup}
\title{List all files in a specified backup Dropbox folder}
\usage{
list_backup(path = path)
}
\arguments{
\item{path}{path to look for files (usually, "commuteR")}
}
\value{
A vector of filenames, without file extensions
}
\description{
List all files in a specified backup Dropbox folder
}
