% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{list_github_packages}
\alias{list_github_packages}
\title{Filter GitHub packages}
\usage{
list_github_packages()
}
\value{
A vector with the name and repo composites, for packages that were installed from GitHub.
}
\description{
Returns a vector with the name and repo composites (e.g. "pinusm/commuteR") for packages that were installed from GitHub.
}
