% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R, R/restore.R
\name{list_packages}
\alias{list_packages}
\title{List all installed packages}
\usage{
list_packages()

list_packages()
}
\value{
A data frame with installed packages names and version numbers

A data frame with the names and versions of all the packages installed
}
\description{
List all installed packages

List all installed packages
}
