% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{populate_backup}
\alias{populate_backup}
\title{Gathers the list of objects to backup and upload to the cloud}
\usage{
populate_backup()
}
\value{
A list of objects to backup and upload to the cloud
}
\description{
Currently, a list of all installed packages, a vector of all CRAN packages, and a vector of all GitHub
    packages, in a repo/packagename composite.
}
