% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{restore_from_cloud}
\alias{restore_from_cloud}
\title{Restore from cloud (Install packages)}
\usage{
restore_from_cloud(path = "commuteR")
}
\arguments{
\item{path}{a folder in Dropbox where the backup files where saved (typically, the same \code{path} given to \code{backup_to_cloud()}}
}
\description{
Restore from cloud (Install packages)
}
